backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.2
pydantic-settings==2.3.1
langchain==0.2.1
langchain-groq
sentence-transformers
chromadb==0.5.0
python-dotenv==1.0.1
alembic==1.13.1
PyPDF2==3.0.1
